Modify data types of amount fields to BigDecimal, and other changes
- Modified data types of amount fields
- Added migrations to reflect the data type change
- Add pre-condition checks for creating topups
- Add pre-condition checks for creating transfers
- Add tests for validators
- Fix tests for transfer calculator
- Clean up schools view
- Fix links/navigation to Parameters
- Add more validation messages

+++ sctp-core/src/main/java/org/cga/sctp/transfers/parameters/EducationTransferParameter.java
@@ -36,6 +36,7 @@ import org.cga.sctp.targeting.importation.parameters.EducationLevel;
 
 import javax.persistence.*;
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Entity
@@ -54,7 +55,7 @@ public class EducationTransferParameter {
     private EducationLevel educationLevel;
 
     @Column
-    private Long amount;
+    private BigDecimal amount;
 
     @Column
     private Boolean active;
@@ -89,11 +90,11 @@ public class EducationTransferParameter {
         this.educationLevel = educationLevel;
     }
 
-    public Long getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
-    public void setAmount(Long amount) {
+    public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
